extract utility functions, move Cart out of main
