+++ src/utils/sortByPrice.jsx
@@ -1,0 +1,9 @@
+export function sortByPrice({ productItems, isSortAscending }) {
+  return productItems.toSorted((a, b) => {
+    if (isSortAscending) {
+      return a.price - b.price;
+    } else {
+      return b.price - a.price;
+    }
+  });
+}
